requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
mysql-connector-python==8.0.26
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
email-validator==1.1.3
Flask-WTF==0.15.1
WTForms==2.3.3
gunicorn==20.1.0
boto3==1.28.38
Pillow==10.0.0 